refactor: extract the VERSION constant into a separate file
also:
- extract VersionInfo into a separate file
- make lib/nokogiri/version.rb require_relative these two new files

so that we can easily find it from Rakefile now,
and potentially a standalone gem spec later.

(This was driven out because Hoe does a questionable thing when
inferring the gem version, which is to grep through files in order of
Manifest.txt; and this is leading to using the LIBXML_REQUIRED_VERSION
as the VERSION. ¯\_(ツ)_/¯)
