+++ lib/nokogiri/version/constant.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+module Nokogiri
+  # The version of Nokogiri you are using
+  VERSION = "1.11.0.rc3"
+end
